Moved handleStateMessage() from WorkflowPresenter to SNAPResponseHandler
