+++ src/snapred/ui/handler/SNAPResponseHandler.py
@@ -57,3 +57,18 @@ class SNAPResponseHandler(QObject):
             )
             messageBox.setDetailedText(f"{message}")
             messageBox.exec()
+
+    def handleStateMessage(self, view):
+        """
+        Handles a specific 'state' message.
+        """
+        from snapred.backend.dao.request.InitializeStateHandler import InitializeStateHandler
+        from snapred.ui.view.InitializeStateCheckView import InitializationMenu
+
+        try:
+            logger.info("Handling 'state' message.")
+            initializationMenu = InitializationMenu(runNumber=InitializeStateHandler.runId, parent=view)
+            initializationMenu.finished.connect(lambda: initializationMenu.deleteLater())
+            initializationMenu.show()
+        except Exception as e:  # noqa: BLE001
+            logger.warning(f"The 'state' handling method encountered an error:{str(e)}")
